Theoretically, should have submodule fixed now.
It seems that assets/extern is set on a branch, which it probably
shouldn't be.
As a result, when I just pulled from main for all submodules,
it tried to pull in changes from main that weren't in the branch.
These then got included in setting the submodule head hash,
which obviously doesn't exist on origin, so we got errors in the
automatic builds. All of this to say, don't leave your submodules on a
branch (which obviously I've done, but I swear I'll change it soon...)